Rename store.js to configureStore.js
