+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
-import configureStore from "./store";
+import configureStore from "./configureStore";
 import "./index.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
